Add bootstrap fix for causal forest fit
